+++ changelog.md
@@ -21,7 +21,12 @@ All notable changes to this project will be documented in this file.
 
 - Create and apply the `@SinceKotoolsAssert` annotation on every declaration
   ([#16](https://github.com/kotools/assert/issues/16)).
-- Add boolean assertions in `kotools.assert`
+- Add boolean and nullability assertions in the `kotools.assert` package
+  ([#15](https://github.com/kotools/assert/issues/15)).
+
+### Deprecated
+
+- Deprecate nullability assertions in the `io.github.kotools.assert` package
   ([#15](https://github.com/kotools/assert/issues/15)).
 
 ## Version 2.0.0 - 2022/04/11